+++ bindgen/src/templates/generic/common/RecordTemplate.kt
@@ -18,7 +18,7 @@ data class {{ type_name }} (
     {%- endif %}
     {% if config.generate_immutable_records() %}val{% else %}var{% endif %} {{ field.name()|var_name }}: {{ field|type_name(ci) -}}
     {%- match field.default_value() %}
-        {%- when Some with(literal) %} = {{ literal|render_default(field, ci) }}
+        {%- when Some with(literal) %} = {{ literal|render_default(field, ci, config) }}
         {%- else %}
         {% if field|is_optional %} = null {% endif %}
     {%- endmatch -%}